% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{auto_format}
\alias{auto_format}
\title{Automatically format community notes data set files using preset transformations}
\usage{
auto_format(x, tz = NULL)
}
\description{
Adds columns for tweet create dates, tweet URLs, and datetime columns for all 'Millis' columns
}
