% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{build_cn_db}
\alias{build_cn_db}
\title{Create a duckdb database using the community notes dataset.}
\usage{
build_cn_db(data_dir, dbdir = data_dir)
}
\arguments{
\item{data_dir}{the directory where the data files are}

\item{dbdir}{the directory where to create the data files}
}
\value{
the connection
}
\description{
Create a duckdb database using the community notes dataset.
}
