% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{cnd_url}
\alias{cnd_url}
\title{Generate the URL to download a Community Notes data set}
\usage{
cnd_url(year, month, day, dataset, number, date = NULL)
}
\description{
Generate the URL to download a Community Notes data set
}
