% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communitynotes-package.R
\docType{package}
\name{communitynotes-package}
\alias{communitynotes}
\alias{communitynotes-package}
\title{communitynotes: Work With The Community Notes Public Datasets}
\description{
Download and read the public community notes datasets and parse the fields.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/communitynotes}
  \item Report bugs at \url{https://github.com/colin-fraser/communitynotes/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
