% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{download_all_data}
\alias{download_all_data}
\title{Download all datasets for a given date}
\usage{
download_all_data(
  date = Sys.Date(),
  download_to = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\description{
Download all datasets for a given date
}
