% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{download_dataset}
\alias{download_dataset}
\title{Download a community notes dataset}
\usage{
download_dataset(
  year,
  month,
  day,
  dataset = c("notes", "ratings", "noteStatusHistory", "userEnrollment"),
  number = 0,
  download_to = NULL,
  increment = FALSE,
  overwrite = FALSE,
  verbose = TRUE,
  is_increment = FALSE
)
}
\description{
Download a community notes dataset
}
