% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{go_to_download_page}
\alias{go_to_download_page}
\title{Open the Commmunity Notes download page in your browser}
\usage{
go_to_download_page()
}
\description{
Open the Commmunity Notes download page in your browser
}
