% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{id_to_datetime}
\alias{id_to_datetime}
\title{Convert a Twitter Snowflake ID to the associated create date}
\usage{
id_to_datetime(x, tz = "UTC")
}
\description{
Convert a Twitter Snowflake ID to the associated create date
}
