% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{latest_data}
\alias{latest_data}
\title{Find the latest data set}
\usage{
latest_data(
  dataset = c("notes", "ratings", "noteStatusHistory", "userEnrollment"),
  start_search = Sys.Date() + 2
)
}
\description{
Find the latest data set
}
