% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{millis_to_datetime}
\alias{millis_to_datetime}
\title{Convert between millisecond timestamp and datetime}
\usage{
millis_to_datetime(x, tz = NULL)
}
\arguments{
\item{x}{For `millis_to_datetime`, the millisecond timestamp to convert.
For `datetime_to_millis`, the datetime object to convert.}

\item{tz}{The time zone to apply to the resulting datetime.
Relevant only for `millis_to_datetime`.}
}
\value{
For `millis_to_datetime`, a datetime object.
        For `datetime_to_millis`, a millisecond timestamp.
}
\description{
Convert between millisecond timestamp and datetime
}
\examples{
millis_to_datetime(1609459200000) # Converts to datetime
datetime_to_millis(as.POSIXct("2021-01-01 UTC")) # Converts to millis
1609459200000 == datetime_to_millis(millis_to_datetime(1609459200000)) # TRUE

}
