% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_and_concat}
\alias{read_and_concat}
\title{Read and Concatenate Community Notes Data Files}
\usage{
read_and_concat(
  directory,
  filetype = c("notes", "ratings", "noteStatusHistory", "userEnrollment"),
  warn_for_filesize = 1e+09
)
}
\arguments{
\item{directory}{The directory from which to read the files.}

\item{filetype}{Type of the files to read, one of "notes", "ratings", "noteStatusHistory", "userEnrollment".}

\item{warn_for_filesize}{Threshold file size in bytes for warning the user.
Defaults to 1e9 (1 GB). Set to Inf to disable warning.}
}
\value{
A data frame with all data from the specified file type concatenated.
}
\description{
Reads all community notes data set files of a specified type from a directory,
concatenating multiple files into a single data frame. You'll be warned if the resulting
file will be quite large. To disable the warning, set warn_for_filesize = Inf.
}
