% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_note_status_history_file}
\alias{read_note_status_history_file}
\title{Read the note status history file}
\usage{
read_note_status_history_file(filename)
}
\description{
Read the note status history file
}
