% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_notes_file}
\alias{read_notes_file}
\title{Read community notes datasets}
\usage{
read_notes_file(filename)
}
\arguments{
\item{filename}{the filename of the file to read}

\item{filetype}{either 'detect' to automatically detect, or one of 'notes', 'ratings', 'userEnrollment'
or 'noteStatusHistory'}
}
\description{
Reads and correctly formats the data files. You probably just want to use read_comunity_notes_file.
}
