% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tweet_url}
\alias{tweet_url}
\title{Generate a URL for a tweetId}
\usage{
tweet_url(x)
}
\arguments{
\item{x}{tweetId}
}
\description{
Note that this won't be the direct URL for the tweet, but it will provide a URL that
you can follow to get to the tweet. This is more of a convenience function for investigation
rather than a canonical URL for the tweet.
}
