% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_and_concat}
\alias{read_and_concat}
\title{Read all the community notes data set files of a certain type out of a directory, concatenating
multiple files together}
\usage{
read_and_concat(
  directory,
  filetype = c("notes", "ratings", "noteStatusHistory", "userEnrollment"),
  warn_for_filesize = 1e+09
)
}
\description{
Read all the community notes data set files of a certain type out of a directory, concatenating
multiple files together
}
